backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Database
supabase==2.3.4
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Spotify API
spotipy==2.23.0

# Romanization Libraries
pypinyin==0.51.0
pykakasi==2.2.1
hangul-romanize==0.1.0

# CORS
python-cors==1.0.0

# Date/Time
python-dateutil==2.8.2

# JSON
orjson==3.9.12

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
